requirements.txt
openai
python-dotenv
discord 
llama_index
qdrant-client
llama-index-vector-stores-qdrant
numpy 
matplotlib
pandas
torch
faiss-cpu